Prevent non-admin ghosts from controlling observed players

+++ code/_onclick/hud/screen_objects.dm
@@ -213,6 +213,9 @@
 	var/last_craft
 
 /atom/movable/screen/craft/Click(location, control, params)
+	if(isobserver(usr) && !IsAdminGhost(usr))
+		return
+
 	var/list/modifiers = params2list(params)
 	if(world.time < lastclick + 3 SECONDS)
 		return
@@ -914,6 +917,9 @@
 		animate(src, alpha = 255, time = 30)
 
 /atom/movable/screen/advsetup/Click(location,control,params)
+	if(isobserver(usr) && !IsAdminGhost(usr))
+		return
+
 	if(!hud)
 		qdel(src)
 		return
@@ -933,6 +939,10 @@
 	icon_state = "eye"
 
 /atom/movable/screen/eye_intent/Click(location,control,params)
+	// Observers borrow the target's HUD while watching them; only admin ghosts may control it.
+	if(isobserver(usr) && !IsAdminGhost(usr))
+		return
+
 	var/list/modifiers = params2list(params)
 	var/_y = text2num(params2list(params)["icon-y"])
 
@@ -2299,6 +2309,9 @@
 
 /atom/movable/screen/read/Click(location, control, params)
 	. = ..()
+	if(isobserver(usr) && !IsAdminGhost(usr))
+		return FALSE
+
 	destroy_read()
 	if(!usr || !usr.client)
 		return FALSE
